rename typesystem.cpp -> core.cpp
